requirements.txt
pandas
numpy
tqdm
python-dotenv
pypdf2
black
